custom hook: useCalculator
